+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -170,6 +170,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - VFX: Preserve specular option for lit outputs (matches HDRP lit shader)
 - Fixed spot light missing from ray tracing indirect effects.
 - Fixed a UI bug in the diffusion profile list after fixing them from the wizard.
+- Fixed the hash collision when creating new diffusion profile assets.
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled